refactor: extract relay-backed terminals; add settings prefix scan
RelayTerminals moves to crates/relay-terminals (zero gpui imports) so
oximux serve exposes the same terminals with the same gap semantics the
desktop does; the desktop keeps only its OnceLock install pattern and
re-exports the shared type. SettingsRepo::list_prefixed adds the one query
the headless session catalog needs — enumerating agent_chat: transcript
blobs — with LIKE metacharacters escaped so a literal %/_ cannot widen the
scan.
